ReadCity/requirements.txt
python-dotenv==1.0.0
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2.1
django-celery==3.1.17
django-cuser==2017.3.16
django-debug-toolbar==5.2.0
django-guardian==3.0.0
django-js-asset==3.1.2
django-modeltranslation==0.19.14
django-mptt==0.17.0
django-rosetta==0.10.2
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
flower==2.0.1
googletrans==4.0.2
guardian==0.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kombu==5.5.3
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
pandas==2.2.3
polib==1.2.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
pymemcache==3.5.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytils==0.4.3
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rosetta==0.3
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tornado==6.5
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary>=2.9
